Add: handle the first line startsWith /
If privacy mode is enable, bot can not access to those messages are not
sent to him, hence, there will be no reply_to_message even though I
reply to a message. The tricky solution is make the replied message
startsWith a /. The bot will know that he should handle this reply.